contentfulの処理をまとめる
